backend/requirements.txt
# Sanoid Manager - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# SSH
paramiko>=3.3.0

# Scheduling
croniter>=2.0.0

# Data Validation
pydantic>=2.5.0
pydantic[email]>=2.5.0

# Environment
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# HTTP Client (per Proxmox API)
aiohttp>=3.9.0

# Email (optional)
aiosmtplib>=3.0.0

# Testing
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
